+++ src/Firebase/Auth/SignIn/GuzzleHandler.php
@@ -190,7 +190,7 @@ final class GuzzleHandler implements Handler
         $body = self::$defaultBody;
 
         if ($action instanceof IsTenantAware && $tenantId = $action->tenantId()) {
-            $body['tenantId'] = $tenantId->toString();
+            $body['tenantId'] = $tenantId;
         }
 
         return $body;