+++ src/Firebase/Auth/SignInWithRefreshToken.php
@@ -6,7 +6,7 @@ namespace Kreait\Firebase\Auth;
 
 final class SignInWithRefreshToken implements IsTenantAware, SignIn
 {
-    private ?TenantId $tenantId = null;
+    private ?string $tenantId = null;
 
     private function __construct(private string $refreshToken)
     {
@@ -17,7 +17,7 @@ final class SignInWithRefreshToken implements IsTenantAware, SignIn
         return new self($refreshToken);
     }
 
-    public function withTenantId(TenantId $tenantId): self
+    public function withTenantId(string $tenantId): self
     {
         $action = clone $this;
         $action->tenantId = $tenantId;
@@ -30,7 +30,7 @@ final class SignInWithRefreshToken implements IsTenantAware, SignIn
         return $this->refreshToken;
     }
 
-    public function tenantId(): ?TenantId
+    public function tenantId(): ?string
     {
         return $this->tenantId;
     }